requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.1
tweepy==4.14.0
boto3==1.28.63